+++ src/core/types/friend.d.ts
@@ -1,6 +1,11 @@
 import { Character } from "./character";
 
-export type Friend = {
+export interface GetFriendWeeklyRaidsRequest {
+  friendUsername: string;
+  characterId: number;
+}
+
+export interface Friend {
   friendId: number;
   friendUsername: string;
   areWeFriend: string;
@@ -8,9 +13,9 @@ export type Friend = {
   characterList: Character[];
   toFriendSettings: FriendSettings;
   fromFriendSettings: FriendSettings;
-};
+}
 
-export type FriendSettings = {
+export interface FriendSettings {
   showDayTodo: boolean;
   showRaid: boolean;
   showWeekTodo: boolean;
@@ -20,12 +25,12 @@ export type FriendSettings = {
   updateGauge: boolean;
   updateRaid: boolean;
   setting: boolean;
-};
+}
 
-export type SearchCharacterItem = {
+export interface SearchCharacterItem {
   id: number;
   areWeFriend: string;
   characterListSize: number;
   characterName: string;
   username: string;
-};
+}
