+++ src/core/utils/queryKeyGenerator.ts
@@ -1,15 +1,18 @@
 import type { QueryKey } from "@tanstack/react-query";
 
-import type { CubeName } from "@core/types/character";
+import type { CubeName, GetWeeklyRaidsRequest } from "@core/types/character";
 import type { GetCommentsRequest } from "@core/types/comment";
+import type { GetFriendWeeklyRaidsRequest } from "@core/types/friend";
 import type { GetNoticeListRequest } from "@core/types/notice";
 
 const defaultKeys = {
   GET_MY_INFORMATION: "GET_MY_INFORMATION",
-  GET_FRIENDS: "GET_FRIENDS",
-  SEARCH_CHARACTER: "SEARCH_CHARACTER",
   GET_CHARACTERS: "GET_CHARACTERS",
+  GET_WEEKLY_RAIDS: "GET_WEEKLY_RAIDS",
   GET_CUBE_REWARD: "GET_CUBE_REWARD",
+  GET_FRIENDS: "GET_FRIENDS",
+  GET_FRIEND_WEEKLY_RAIDS: "GET_FRIEND_WEEKLY_RAIDS",
+  SEARCH_CHARACTER: "SEARCH_CHARACTER",
   GET_COMMENTS: "GET_COMMENTS",
   GET_OFFICIAL_NOTICES: "GET_OFFICIAL_NOTICES",
   GET_NOTICES: "GET_NOTICES",
@@ -27,17 +30,23 @@ const queryKeyGenerator = {
   getMyInformation: () => {
     return withParamGenerator(defaultKeys.GET_MY_INFORMATION);
   },
-  getFriends: () => {
-    return withParamGenerator(defaultKeys.GET_FRIENDS);
-  },
   searchCharacter: (characterName?: string) => {
     return withParamGenerator(defaultKeys.SEARCH_CHARACTER, characterName);
   },
   getCharacters: () => {
     return withParamGenerator(defaultKeys.GET_CHARACTERS);
   },
+  getWeeklyRaids: (params?: GetWeeklyRaidsRequest) => {
+    return withParamGenerator(defaultKeys.GET_WEEKLY_RAIDS, params);
+  },
   getCubeReward: (name?: CubeName) => {
     return withParamGenerator(defaultKeys.GET_CUBE_REWARD, name);
+  },
+  getFriends: () => {
+    return withParamGenerator(defaultKeys.GET_FRIENDS);
+  },
+  getFriendWeeklyRaid: (params?: GetFriendWeeklyRaidsRequest) => {
+    return withParamGenerator(defaultKeys.GET_FRIEND_WEEKLY_RAIDS, params);
   },
   getComments: (params?: GetCommentsRequest) => {
     return withParamGenerator(defaultKeys.GET_COMMENTS, params);
